captouch: petal 5 positional data fixed
This raises CPU load of captouch from 3% to 7%.
We think we golfed i2c to the minimum but it's
still a lot of blocking calls. esp-idf 5.2
provides a sort-of nonblocking i2c lib, so this
can be improved if we bump the version.

Cycle time for both chips is 18ms as before,
this can be raised to 9ms(top)/14ms(bottom) but
it increases the CPU load to 9% so we'll wait
with this until we're not busyspinning anymore.

This also deprecates the "pressed" detection
for individual petal pads for forward compat
reasons, it always returns False for now.
Use the overall petal pressed state plus
position detection instead.